==> extra/one_gau_cluster/2_gen_best_model_from_pacmap.R <==
library(quollr)
library(dplyr)
library(readr)
library(ggplot2)

## Import data
training_data_gau <- read_rds("data/one_gau_cluster/data_one_gau_training.rds")

pacmap_gau <- read_rds("data/one_gau_cluster/pacmap_data_one_gau.rds")
gau_scaled_obj <- gen_scaled_data(
  data = pacmap_gau)
pacmap_gau_scaled <- gau_scaled_obj$scaled_nldr

## To initialise number of bins along the x-axis
bin1_vec <- 2:25
buff_vec <- seq(0.05, 0.2, by = 0.01)

lim1 <- gau_scaled_obj$lim1
lim2 <- gau_scaled_obj$lim2
r2_gau <- diff(lim2)/diff(lim1)

error_gau <- data.frame(matrix(nrow = 0, ncol = 0))

for (xbins in bin1_vec) {
  for(q in buff_vec) {

    bin2 <- calc_bins_y(bin1 = xbins, r2 = r2_gau, q = q)$bin2

    gau_model <- fit_highd_model(
      training_data = training_data_gau,
      emb_df = pacmap_gau_scaled,
      bin1 = xbins,
      r2 = r2_gau,
      is_bin_centroid = TRUE,
      is_rm_lwd_hex = FALSE,
      col_start_highd = "x",
      q = q
    )

    df_bin_centroids_gau <- gau_model$df_bin_centroids
    df_bin_gau <- gau_model$df_bin

    ## Compute error
    error_df <- glance(
      df_bin_centroids = df_bin_centroids_gau,
      df_bin = df_bin_gau,
      training_data = training_data_gau,
      newdata = NULL,
      type_NLDR = "PaCMAP",
      col_start = "x") |>
      mutate(bin1 = xbins,
             bin2 = bin2,
             b = bin1 * bin2,
             b_non_empty = NROW(df_bin_centroids_gau),
             q = q)

    error_gau <- bind_rows(error_gau, error_df)

  }

}

error_gau |>
  arrange(MSE)


ggplot(error_gau, aes(x = b_non_empty,
                      y = log(MSE))) +
  geom_point() +
  geom_line() +
  geom_vline(xintercept = 65, linetype="solid",
             color = "black", linewidth=0.8, alpha = 0.5)
